extract OpenAI client as a separate module & add `max_tokens` setting
