+++ src/app.rs
@@ -15,7 +15,7 @@ use crate::{
     database::{DatabaseManager, FileDatabaseProvider, InMemDatabaseProvider},
     dispatcher::build_dispatcher,
     module_mgr::ModuleManager,
-    modules::{admin::Admin, chat::Chat, prefs::Prefs, stats::Stats},
+    modules::{admin::Admin, chat::Chat, openai::OpenAI, prefs::Prefs, stats::Stats},
     types::HandlerResult,
 };
 
@@ -55,6 +55,7 @@ pub async fn run(config: SharedConfig) {
     debug!("Initializing modules...");
     let mut module_mgr = ModuleManager::new();
     module_mgr.register_module(crate::modules::config::Config::new(config.clone()));
+    module_mgr.register_module(OpenAI);
     module_mgr.register_module(Prefs::new(db_mgr.clone()));
     module_mgr.register_module(Admin::new(db_mgr.clone()));
     module_mgr.register_module(Stats::new(db_mgr.clone()));
