+++ src/modules/mod.rs
@@ -3,5 +3,6 @@
 pub(crate) mod admin;
 pub(crate) mod chat;
 pub(crate) mod config;
+pub(crate) mod openai;
 pub(crate) mod prefs;
 pub(crate) mod stats;
